Move backround update login to separate class
